Seperate schema from models
