+++ packages/server/models/event.model.js
@@ -1,19 +1,4 @@
 const mongoose = require("mongoose");
-const uuid = require("uuid");
-
-const EventSchema = mongoose.Schema({
-  _id: {
-    type: String,
-    default: uuid.v4
-  },
-  name: {
-    type: String,
-    required: true
-  },
-  slug: {
-    type: String,
-    required: true
-  }
-});
+const { EventSchema } = require("./schema");
 
 module.exports = mongoose.model("Event", EventSchema);
