+++ packages/server/models/schema/index.js
@@ -1,0 +1,17 @@
+const AdminSchema = require("./admin.schema");
+const ApplicationSchema = require("./application.schema");
+const EventSchema = require("./event.schema");
+const HackerSchema = require("./hacker.schema");
+const ReviewSchema = require("./review.schema");
+const SettingsSchema = require("./settings.schema");
+const UserSchema = require("./user.schema");
+
+module.exports = {
+  AdminSchema,
+  ApplicationSchema,
+  EventSchema,
+  HackerSchema,
+  ReviewSchema,
+  SettingsSchema,
+  UserSchema
+};
