+++ packages/server/models/schema/settings.schema.js
@@ -1,0 +1,15 @@
+const mongoose = require("mongoose");
+const uuid = require("uuid");
+
+const SettingsSchema = mongoose.Schema({
+  _id: {
+    type: String,
+    default: uuid.v4
+  },
+  numberOfReviewsRequired: {
+    type: Number,
+    required: true
+  }
+});
+
+module.exports = SettingsSchema;
