+++ packages/server/models/settings.model.js
@@ -1,15 +1,4 @@
 const mongoose = require("mongoose");
-const uuid = require("uuid");
-
-const SettingsSchema = mongoose.Schema({
-  _id: {
-    type: String,
-    default: uuid.v4
-  },
-  numberOfReviewsRequired: {
-    type: Number,
-    required: true
-  }
-});
+const { SettingsSchema } = require("./schema");
 
 module.exports = mongoose.model("Settings", SettingsSchema);
